+++ src/main/java/com/cos/bank/config/dummy/DummyObject.java
@@ -8,6 +8,7 @@ import com.cos.bank.user.domain.Role;
 import com.cos.bank.user.domain.User;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class DummyObject {
@@ -52,12 +53,12 @@ public class DummyObject {
         return Account.builder()
                 .number(number)
                 .password(encodedPw)
-                .balance(100.0)
+                .balance(BigDecimal.valueOf(100.0))
                 .user(user)
                 .build();
     }
 
-    protected Account newMockAccount(Long id, Long number, Double balance, User user) {
+    protected Account newMockAccount(Long id, Long number, BigDecimal balance, User user) {
         String encodedPw = passwordEncoder.encode("1234");
 
         return Account.builder()
@@ -72,12 +73,12 @@ public class DummyObject {
     }
 
     protected Transaction newMockDepositTransaction(Long id, Account account){
-        account.deposit(10.0);
+        account.deposit(BigDecimal.valueOf(10.0));
         return Transaction.builder()
                 .id(id)
                 .withdrawAccount(null)
                 .depositAccount(account)
-                .amount(10.0)
+                .amount(BigDecimal.valueOf(10.0))
                 .withdrawAccountBalance(null)
                 .depositAccountBalance(account.getBalance())
                 .transactionType(TransactionType.DEPOSIT)
@@ -90,12 +91,12 @@ public class DummyObject {
     }
 
     protected Transaction newMockWithdrawTransaction(Long id, Account account){
-        account.withdraw( 10.0);
+        account.withdraw(BigDecimal.valueOf(10.0));
         return Transaction.builder()
                 .id(id)
                 .withdrawAccount(account)
                 .depositAccount(null)
-                .amount(10.0)
+                .amount(BigDecimal.valueOf(10.0))
                 .withdrawAccountBalance(account.getBalance())
                 .depositAccountBalance(null)
                 .transactionType(TransactionType.WITHDRAW)
@@ -107,7 +108,7 @@ public class DummyObject {
     }
 
     protected Transaction newWithdrawTransaction(Account account, AccountRepository accountRepository) {
-        account.withdraw(10.0);
+        account.withdraw(BigDecimal.valueOf(10.0));
 
         // Repository Test - dirty checking (o)
         // Controller Test - dirty checking (x)
@@ -117,7 +118,7 @@ public class DummyObject {
         return Transaction.builder()
                 .withdrawAccount(account)
                 .depositAccount(null)
-                .amount(10.0)
+                .amount(BigDecimal.valueOf(10.0))
                 .withdrawAccountBalance(account.getBalance())
                 .depositAccountBalance(null)
                 .transactionType(TransactionType.WITHDRAW)
@@ -129,7 +130,7 @@ public class DummyObject {
     }
 
     protected Transaction newDepositTransaction(Account account, AccountRepository accountRepository) {
-        account.deposit(10.0);
+        account.deposit(BigDecimal.valueOf(10.0));
 
         // Repository Test - dirty checking (o)
         // Controller Test - dirty checking (x)
@@ -140,7 +141,7 @@ public class DummyObject {
         return Transaction.builder()
                 .withdrawAccount(null)
                 .depositAccount(account)
-                .amount(10.0)
+                .amount(BigDecimal.valueOf(10.0))
                 .withdrawAccountBalance(null)
                 .depositAccountBalance(account.getBalance())
                 .transactionType(TransactionType.DEPOSIT)
@@ -154,8 +155,8 @@ public class DummyObject {
 
     protected Transaction newTransferTransaction(Account withdrawAccount, Account depositAccount,
                                                  AccountRepository accountRepository) {
-        withdrawAccount.withdraw(10.0);
-        depositAccount.deposit(10.0);
+        withdrawAccount.withdraw(BigDecimal.valueOf(10.0));
+        depositAccount.deposit(BigDecimal.valueOf(10.0));
 
         // Repository Test - dirty checking (o)
         // Controller Test - dirty checking (x)
@@ -169,7 +170,7 @@ public class DummyObject {
                 .depositAccount(depositAccount)
                 .withdrawAccountBalance(withdrawAccount.getBalance())
                 .depositAccountBalance(depositAccount.getBalance())
-                .amount(10.0)
+                .amount(BigDecimal.valueOf(10.0))
                 .transactionType(TransactionType.TRANSFER)
                 .sender(String.valueOf(withdrawAccount.getNumber()))
                 .receiver(String.valueOf(depositAccount.getNumber()))
